copy []byte before returning
The []byte value bolt returns points directly to the
underlying database page and can not be retained after
closing the transaction. This is a sharp edge and all
other values we return are already copies of the
underlying data. Copying here makes it safe to use the
returned value in any way.

If the client needs to avoid this copy for performance
reasons, they can bypass genbolt for that value and read
it directly from the bucket.