factor out godegen template and schema-schema
This is the result of thinking more about the metaschema,
the internal schema that describes the user's schema. It's
now consolidated into one place (schema.go) and
significantly streamlined. It is hopefully now a reasonably
clear description of the actual structure of a genbolt
database schema, with each part having an obvious purpose,
and with fewer incidental pieces that are just artifacts of
the internals of genbolt.

Part of that is taking all the ad hoc codegen (the printfs
and templates for various fragments of the output) and
combining them into one big template. This makes it easier
to find things.