rely on go/types where possible
This uses go/types rather than go/ast for more logic.
This makes comparisons and transformations more precise.
(For example, a user-defined type named "bool" won't be
confused with Go's basic type bool.) This also makes the
metaschema a little more typesafe and regular. (We've
removed the one remaining field that was defined as
"either a *types.Pointer or a string", now it's just
types.Type.)