add test flag to save test output
This is useful when making a deliberate change to the format
of the generated code. Instead of updating all the output files
by hand, you can use genbolt to do it automatically. Care must
be taken to avoid introducing bugs this way; use this flag only
when you have high confidence the change is safe, and only for
changes that would take a lot of work to do by hand.