gracefully handle nonexistent buckets in read-only txs
Our general approach is to make the database act as if
the entire schema exists with zero values from the
beginning. Buckets are created on demand. But if a
bucket hasn't been created yet, we should behave as if
it exists and is empty (and recursively for its child
buckets and records).